requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
jsonschema==3.0.1
Pillow==5.4.1
pyrsistent==0.15.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
